+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG
 ### Version 2.1.2
 - ADD: Making formula page aSync
+- FIX: Formula not showning properly when new
 
 ### Version 2.1.1
 - ADD: Privacy notice for the PV Online services